renamed UI components (root cause analysis -> analysis)
