+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/analysis/analysis.html
@@ -1,15 +1,15 @@
-<template name="RCA_Analysis_Edit">
+<template name="Analysis_Edit">
   <div class="card-block">
     {{#with analysis}}
       {{#FormField label=label}}{{/FormField}}
 
-      {{> RCA_TargetDate_Edit
+      {{> Analysis_TargetDate_Edit
           date=this.targetDate
           onUpdate=methods.setDate
           disabled=isAnalysisCompleted}}
 
       {{#unless isAnalysisCompleted}}
-        {{> RCA_ToBeCompletedBy_Edit
+        {{> Analysis_ToBeCompletedBy_Edit
             assignee=this.executor
             comments=this.completionComments
             isButtonVisible=true
@@ -18,14 +18,14 @@
             onUpdate=methods.setExecutor
             onComplete=methods.complete}}
       {{else}}
-        {{> RCA_TargetDate_Edit
+        {{> Analysis_TargetDate_Edit
             date=this.completedAt
             placeholder="Completed date"
             label="Completed date"
             disabled=areCompletionFieldsDisabled
             onUpdate=methods.setCompletedDate}}
 
-        {{> RCA_ToBeCompletedBy_Edit
+        {{> Analysis_ToBeCompletedBy_Edit
             assignee=this.completedBy
             isButtonVisible=false
             canButtonBeShown=(not areStandardsUpdated)
@@ -35,7 +35,7 @@
             onUpdate=methods.setCompletedBy
             onUndo=methods.undo}}
 
-        {{> RCA_Comments_Edit
+        {{> Analysis_Comments_Edit
             comments=this.completionComments
             disabled=areCompletionFieldsDisabled
             onUpdate=methods.setComments}}
