+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/analysis/analysis.js
@@ -1,6 +1,6 @@
 import { Template } from 'meteor/templating';
 
-Template.RCA_Analysis_Edit.viewmodel({
+Template.Analysis_Edit.viewmodel({
   mixin: 'utils',
   label: 'Root cause analysis',
   analysis: '',
