+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/includes/assignee/assignee.html
@@ -1,4 +1,4 @@
-<template name="RCA_ToBeCompletedBy_Edit">
+<template name="Analysis_ToBeCompletedBy_Edit">
   {{#FormField label=label}}
     {{#Completion_Edit
       isButtonVisible=isButtonVisible
