+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/includes/assignee/assignee.js
@@ -1,6 +1,6 @@
 import { Template } from 'meteor/templating';
 
-Template.RCA_ToBeCompletedBy_Edit.viewmodel({
+Template.Analysis_ToBeCompletedBy_Edit.viewmodel({
   assignee() { return Meteor.userId() },
   comments: '',
   selectFirstIfNoSelected: false,
