+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/includes/comments/comments.html
@@ -1,4 +1,4 @@
-<template name="RCA_Comments_Edit">
+<template name="Analysis_Comments_Edit">
   {{#FormField label="Comments"}}
     <textarea class="form-control" rows="3"
         {{b "value: comments, focusout: update, placeholder: placeholder, disable: disabled"}}></textarea>
