+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/includes/date/date.html
@@ -1,4 +1,4 @@
-<template name="RCA_TargetDate_Edit">
+<template name="Analysis_TargetDate_Edit">
   {{#FormField label=label}}
     {{> Datepicker
         startDate=startDate
