+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/update-of-standards/update-of-standards.html
@@ -1,16 +1,16 @@
-<template name="RCA_UOS_Edit">
+<template name="UOS_Edit">
   <div class="card-block">
     {{#with updateOfStandards}}
       {{#FormField label=label}}{{/FormField}}
 
-      {{> RCA_TargetDate_Edit
+      {{> Analysis_TargetDate_Edit
           startDate=defaultTargetDate
           date=this.targetDate
           onUpdate=methods.setDate
           disabled=this.status}}
 
       {{#unless this.status}}
-        {{> RCA_ToBeCompletedBy_Edit
+        {{> Analysis_ToBeCompletedBy_Edit
             assignee=this.executor
             comments=this.completionComments
             isButtonVisible=true
@@ -19,14 +19,14 @@
             onUpdate=methods.setExecutor
             onComplete=methods.complete}}
       {{else}}
-        {{> RCA_TargetDate_Edit
+        {{> Analysis_TargetDate_Edit
             date=this.completedAt
             placeholder="Completed date"
             label="Completed date"
             disabled=areCompletionFieldsDisabled
             onUpdate=methods.setCompletedDate}}
 
-        {{> RCA_ToBeCompletedBy_Edit
+        {{> Analysis_ToBeCompletedBy_Edit
             assignee=this.completedBy
             isButtonVisible=false
             placeholder="Completed by"
@@ -35,7 +35,7 @@
             onUpdate=methods.setCompletedBy
             onUndo=methods.undo}}
 
-        {{> RCA_Comments_Edit
+        {{> Analysis_Comments_Edit
             comments=this.completionComments
             disabled=areCompletionFieldsDisabled
             onUpdate=methods.setComments}}
