+++ app/imports/ui/components/includes/subcards/analysis-wrapper/edit/update-of-standards/update-of-standards.js
@@ -1,6 +1,6 @@
 import { Template } from 'meteor/templating';
 
-Template.RCA_UOS_Edit.viewmodel({
+Template.UOS_Edit.viewmodel({
   mixin: 'utils',
   label: 'Update of standard(s)',
   updateOfStandards: '',
