+++ app/imports/ui/components/includes/subcards/analysis-wrapper/read/analysis-wrapper.js
@@ -1,7 +1,7 @@
 import { Template } from 'meteor/templating';
 import { AnalysisTitles } from '/imports/api/constants.js';
 
-Template.Subcards_RootCauseAnalysis_Read.viewmodel({
+Template.Subcards_AnalysisWrapper_Read.viewmodel({
   RCALabel: AnalysisTitles.rootCauseAnalysis,
   UOSLabel: 'Update of standard(s)'
 });
