+++ app/imports/ui/components/includes/subcards/analysis-wrapper/read/index.js
@@ -1,0 +1,3 @@
+import './analysis-wrapper.html';
+import './analysis-wrapper.js';
+import './item';
