+++ app/imports/ui/components/includes/subcards/analysis-wrapper/read/item/item.html
@@ -1,4 +1,4 @@
-<template name="Subcards_RootCauseAnalysis_Read_Item">
+<template name="Subcards_Analysis_Read_Item">
   {{#with doc}}
     {{#if every this.executor this.targetDate}}
       <div class="card-block card-subheading">
