+++ app/imports/ui/components/includes/subcards/analysis-wrapper/read/item/item.js
@@ -2,7 +2,7 @@ import { Template } from 'meteor/templating';
 
 import { AnalysisStatuses } from '/imports/api/constants.js';
 
-Template.Subcards_RootCauseAnalysis_Read_Item.viewmodel({
+Template.Subcards_Analysis_Read_Item.viewmodel({
   mixin: ['user', 'date', 'utils'],
   getAnalysisStatusName(status = 0) {
     return AnalysisStatuses[status];
