+++ app/imports/ui/components/includes/subcards/index.js
@@ -3,7 +3,7 @@ import './improvement-plan';
 import './lessons-learned';
 import './notify';
 import './non-conformities';
-import './root-cause-analysis';
+import './analysis-wrapper';
 import './occurrences';
 import './other-files';
 import './risk-scoring';
