--- app/imports/ui/components/includes/subcards/root-cause-analysis/read/index.js
@@ -1,3 +1,0 @@
-import './root-cause-analysis.html';
-import './root-cause-analysis.js';
-import './item';
