+++ app/imports/ui/components/non-conformities/card/card.html
@@ -95,7 +95,7 @@
 
           </div>
 
-          {{> Subcards_RootCauseAnalysis_Read
+          {{> Subcards_AnalysisWrapper_Read
               analysis=this.analysis
               updateOfStandards=this.updateOfStandards}}
