+++ app/imports/ui/components/non-conformities/includes/edit/main/main.html
@@ -46,7 +46,7 @@
     </div>
 
     {{#if showRootCauseAnalysis}}
-      {{> Subcards_RCA_Edit (RCAArgs this)}}
+      {{> Subcards_AnalysisWrapper_Edit (RCAArgs this)}}
     {{/if}}
   {{/with}}
 </template>
