+++ app/imports/ui/components/risks/card/card.html
@@ -75,7 +75,7 @@
 
           </div>
 
-          {{> Subcards_RootCauseAnalysis_Read
+          {{> Subcards_AnalysisWrapper_Read
               analysis=this.analysis
               updateOfStandards=this.updateOfStandards
               RCALabel=RiskRCALabel
