+++ app/imports/ui/components/risks/includes/edit/edit.html
@@ -7,7 +7,7 @@
             methodRefs=getMethodRefs}}
 
         {{#if showRootCauseAnalysis}}
-          {{> Subcards_RCA_Edit
+          {{> Subcards_AnalysisWrapper_Edit
               _id=this._id
               analysis=this.analysis
               updateOfStandards=this.updateOfStandards
