+++ app/imports/ui/components/standard-subcards/card/card.html
@@ -8,7 +8,7 @@
       {{! NON-CONFORMITIES SUBCARDS}}
 
       {{#with NC}}
-        {{> Subcards_RootCauseAnalysis_Read
+        {{> Subcards_AnalysisWrapper_Read
           analysis=this.analysis
           updateOfStandards=this.updateOfStandards}}
         {{#with notify}}
